+++ editor/editor-core/src/__tests__/playwright/media/hyperlink-toolbar.spec.ts
@@ -62,25 +62,28 @@ test.describe('hyperlink toolbar', () => {
     await expect(editor).toMatchDocument(doc(p('Click Me'), p()));
   });
 
-  test('inserts a link when tabbing through hyperlink toolbar', async ({
-    editor,
-  }) => {
-    const smartLinkToolbarModel = EditorLinkFloatingToolbarModel.from(editor);
+  // DTR-1554 Clear button is not selectable on Safari by Tab
+  test.fixme(
+    'inserts a link when tabbing through hyperlink toolbar',
+    async ({ editor }) => {
+      const smartLinkToolbarModel = EditorLinkFloatingToolbarModel.from(editor);
 
-    await smartLinkToolbarModel.openViaKeyboardShortcut();
+      await smartLinkToolbarModel.openViaKeyboardShortcut();
 
-    await editor.keyboard.type('www.atlassian.com');
-    await editor.keyboard.press('Tab'); // To clear link button
-    await editor.keyboard.press('Tab'); // To label field
-    await editor.keyboard.type('Hello world!');
-    await editor.keyboard.press('Tab'); // To clear text button
-    await editor.keyboard.press('Tab'); // Submit
+      await editor.page.pause();
+      await editor.keyboard.type('www.atlassian.com');
+      await editor.keyboard.press('Tab'); // To clear link button
+      await editor.keyboard.press('Tab'); // To label field
+      await editor.keyboard.type('Hello world!');
+      await editor.keyboard.press('Tab'); // To clear text button
+      await editor.keyboard.press('Tab'); // Submit
 
-    const smartLinkModel = EditorLinkModel.from(editor);
-    await smartLinkModel.isVisibleByText('Hello world!');
+      const smartLinkModel = EditorLinkModel.from(editor);
+      await smartLinkModel.isVisibleByText('Hello world!');
 
-    await expect(editor).toMatchDocument(
-      doc(p(a({ href: 'http://www.atlassian.com' })('Hello world!'))),
-    );
-  });
+      await expect(editor).toMatchDocument(
+        doc(p(a({ href: 'http://www.atlassian.com' })('Hello world!'))),
+      );
+    },
+  );
 });