+++ editor/editor-core/src/__tests__/playwright/table/__fixtures__/resize-documents.ts
@@ -1682,5 +1682,5 @@ export const createSquareTable = ({
     return tr(...tableRowContent);
   });
 
-  return doc(table()(...tableContent));
+  return doc(table({ localId: 'localId' })(...tableContent));
 };