+++ media/media-picker/.babelrc.js
@@ -1,0 +1,7 @@
+const numericSeparatorPreset = {
+  plugins: [['@babel/plugin-proposal-numeric-separator']],
+};
+
+module.exports = {
+  presets: [numericSeparatorPreset],
+};