+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 183.1.7
+
+### Patch Changes
+
+- Updated dependencies
+
+## 183.1.6
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 183.1.5
 
 ### Patch Changes